Removed old 'GE'
Closes #59, #60
