--- MarioDoubleEvolution/src/main/kotlin/cz/cuni/mff/aspect/evolution/levels/ge/algorithm/GenesToSentenceConverter.kt
@@ -1,45 +1,0 @@
-package cz.cuni.mff.aspect.evolution.levels.ge.algorithm
-
-import kotlin.math.abs
-
-
-class GenesToSentenceConverter(private val grammar: Grammar) {
-
-    fun convert(genes: ByteArray, maxWrapsCount: Int = 100): GrammarSentence {
-        val unsignedGenes = genes.map { it.toInt() + abs(Byte.MIN_VALUE.toInt()) }.toTypedArray()
-        return this.convertUnsigned(unsignedGenes, maxWrapsCount)
-    }
-
-    fun convertUnsigned(unsignedGenes: Array<Int>, maxWrapsCount: Int = 100): GrammarSentence {
-        val currentSentence = mutableListOf<Symbol>(this.grammar.startingSymbol)
-
-        var firstNonTerminal = currentSentence.find { it.expandable }
-        val genesIterator = CircularIterator(unsignedGenes)
-
-        while (firstNonTerminal != null && genesIterator.wrapsCount < maxWrapsCount) {
-            val ruleIndex: Int = genesIterator.next()
-            val ruleCandidates = this.grammar.getRules(firstNonTerminal)
-            val ruleToUse = ruleCandidates[ruleIndex % ruleCandidates.size]
-
-            val firstNonTerminalIndex = currentSentence.indexOf(firstNonTerminal)
-            currentSentence.removeAt(firstNonTerminalIndex)
-
-            val rightSideOfRuleToUse = ruleToUse.to.map { s -> s.copy() }
-            rightSideOfRuleToUse.forEach { it.takeParameters(genesIterator) }
-            currentSentence.addAll(firstNonTerminalIndex, rightSideOfRuleToUse)
-
-            firstNonTerminal = currentSentence.find { it.expandable }
-        }
-
-        if (genesIterator.wrapsCount > 0) println("Wraps count: ${genesIterator.wrapsCount}")
-        if (genesIterator.wrapsCount == maxWrapsCount) {
-            return emptyArray()
-        }
-
-        return currentSentence.toTypedArray()
-    }
-
-}
-
-
-typealias GrammarSentence = Array<Symbol>
