+++ MarioDoubleEvolution/src/main/kotlin/cz/cuni/mff/aspect/evolution/levels/pc/chunks/MarioLevelChunk.kt
@@ -1,4 +1,4 @@
-package cz.cuni.mff.aspect.evolution.levels.chunks.chunks
+package cz.cuni.mff.aspect.evolution.levels.pc.chunks
 
 import java.io.Serializable
