+++ MarioDoubleEvolution/src/main/kotlin/cz/cuni/mff/aspect/evolution/levels/pc/evaluators/DistanceLinearityDifficultyCompressionEvaluator.kt
@@ -1,12 +1,11 @@
-package cz.cuni.mff.aspect.evolution.levels.chunks.evaluators
+package cz.cuni.mff.aspect.evolution.levels.pc.evaluators
 
-import cz.cuni.mff.aspect.evolution.levels.chunks.metadata.ChunksLevelMetadata
+import cz.cuni.mff.aspect.evolution.levels.pc.metadata.ChunksLevelMetadata
 import cz.cuni.mff.aspect.mario.GameStatistics
 import cz.cuni.mff.aspect.mario.level.MarioLevel
-import cz.cuni.mff.aspect.utils.discretize
 import io.jenetics.Optimize
 
-class DistanceLinearityDifficultyCompressionDiscretizedEvaluator : SummingEvaluator() {
+class DistanceLinearityDifficultyCompressionEvaluator : SummingEvaluator() {
 
     override fun evaluateOne(level: MarioLevel, levelMetadata: ChunksLevelMetadata, gameStatistics: GameStatistics): Float {
         val distance = gameStatistics.finalMarioDistance
@@ -15,11 +14,7 @@ class DistanceLinearityDifficultyCompressionDiscretizedEvaluator : SummingEvalua
         val linearityFactor = LinearityEvaluator().evaluateOne(level, levelMetadata, gameStatistics)
         val compressionFactor = HuffmanCompressionEvaluator().evaluateOne(level, levelMetadata, gameStatistics)
 
-        val linearityDiscretized = discretize(linearityFactor, arrayOf(0.0f, 0.3f, 0.6f, 1.0f))
-        val difficultyDiscretized = discretize(difficultyFactor, arrayOf(0.0f, 0.3f, 0.6f, 1.0f))
-        val compressionDiscretized = discretize(compressionFactor, arrayOf(0.0f, 0.3f, 0.6f, 1.0f))
-
-        return distance * (1 + linearityDiscretized + difficultyDiscretized + compressionDiscretized)
+        return distance * (1 + difficultyFactor + linearityFactor + compressionFactor)
     }
 
     override val optimize: Optimize get() = Optimize.MAXIMUM
