+++ MarioDoubleEvolution/src/main/kotlin/cz/cuni/mff/aspect/launch/EvolveLGMany.kt
@@ -2,7 +2,6 @@ package cz.cuni.mff.aspect.launch
 
 import ch.idsia.agents.IAgent
 import cz.cuni.mff.aspect.evolution.levels.LevelPostProcessor
-import cz.cuni.mff.aspect.evolution.levels.ge.GrammarLevelEvolution
 import cz.cuni.mff.aspect.evolution.levels.pmp.PMPLevelGeneratorEvolution
 import cz.cuni.mff.aspect.evolution.levels.pmp.evaluators.DistanceLinearityDifficultyCompressionDiscretizedEvaluator
 import cz.cuni.mff.aspect.evolution.levels.pmp.evaluators.PMPLevelGeneratorEvaluator
@@ -11,7 +10,6 @@ import cz.cuni.mff.aspect.storage.LevelStorage
 import cz.cuni.mff.aspect.visualisation.level.LevelVisualiser
 
 fun main() {
-//    doManyGrammarEvolution()
     doManyPMPEvolution()
 }
 
@@ -78,22 +76,6 @@ fun doManyPMPEvolution() {
 }
 
 
-fun doManyGrammarEvolution() {
-
-    val launchers = arrayOf(
-        GrammarEvolutionLauncher(
-            label = "firstManyTest",
-            agentFactory = { Agents.NeuroEvolution.BestGeneric },
-            populationSize = 50,
-            generationsCount = 50,
-            levelsCount = 1,
-            postProcess = false
-        )
-    )
-
-    launchers.forEach { it.launch() }
-}
-
 class PMPEvolutionLauncher(
     private val storageLocation: String,
     private val label: String,
@@ -127,33 +109,4 @@ class PMPEvolutionLauncher(
         levels.forEachIndexed { index, level -> levelVisualiser.store(level, "data/levels/experiments/$storageLocation/${label}_$index") }
         levelEvolution.chart.store("data/levels/experiments/$storageLocation/$label")
     }
-}
-
-class GrammarEvolutionLauncher(
-    private val label: String,
-    private val agentFactory: () -> IAgent,
-    private val populationSize: Int,
-    private val generationsCount: Long,
-    private val levelsCount: Int,
-    private val postProcess: Boolean
-) {
-
-    private val levelVisualiser = LevelVisualiser()
-
-    fun launch()  {
-        val levelEvolution = GrammarLevelEvolution(
-            levelsCount = this.levelsCount,
-            populationSize = this.populationSize,
-            generationsCount = this.generationsCount
-        )
-
-        var levels = levelEvolution.evolve(this.agentFactory)
-
-        if (this.postProcess)
-            levels = levels.map { LevelPostProcessor.postProcess(it) }.toTypedArray()
-
-        levels.forEachIndexed { index, level -> LevelStorage.storeLevel("data/levels/experiments/$label/$index.lvl", level) }
-        levels.forEachIndexed { index, level -> levelVisualiser.store(level, "data/levels/experiments/$label/$index") }
-    }
-
 }
