+++ MarioDoubleEvolution/src/main/kotlin/cz/cuni/mff/aspect/launch/PlayMarioKeyboard.kt
@@ -2,9 +2,7 @@ package cz.cuni.mff.aspect.launch
 
 import ch.idsia.agents.controllers.keyboard.CheaterKeyboardAgent
 import cz.cuni.mff.aspect.evolution.levels.LevelPostProcessor
-import cz.cuni.mff.aspect.evolution.levels.chunks.PCLevelGenerator
-import cz.cuni.mff.aspect.evolution.results.Agents
-import cz.cuni.mff.aspect.evolution.results.LevelGenerators
+import cz.cuni.mff.aspect.evolution.levels.pc.PCLevelGenerator
 import cz.cuni.mff.aspect.mario.GameSimulator
 import cz.cuni.mff.aspect.mario.level.MarioLevel
 import cz.cuni.mff.aspect.visualisation.level.LevelVisualiser
