+++ MarioDoubleEvolution/src/main/kotlin/cz/cuni/mff/aspect/mario/level/original/Stage5Level1.kt
@@ -1,6 +1,6 @@
 package cz.cuni.mff.aspect.mario.level.original
 
-import cz.cuni.mff.aspect.evolution.levels.ge.grammar.ColumnHelpers
+import cz.cuni.mff.aspect.evolution.levels.pc.chunks.ColumnHelpers
 import cz.cuni.mff.aspect.mario.Entities
 import cz.cuni.mff.aspect.mario.Tiles
 import cz.cuni.mff.aspect.mario.level.MarioLevel
